+++ ruma-client-api/CHANGELOG.md
@@ -8,7 +8,8 @@ Bug fixes:
 
 Breaking changes:
 
-* Update type of `contains_url` in `r0::filter::RoomEventFilter` from `Option<bool>` to `Option<ContainsUrlFilter>`.
+* Update `contains_url: Option<bool>` in `r0::filter::RoomEventFilter` to
+  `url_filter: Option<UrlFilter>`.
 * Borrow strings in outgoing requests and responses.
   * Explicit types may have to be updated from `endpoint::Request` to `endpoint::Request<'_>` on
     clients and `endpoint::IncomingRequest` on servers, the other way around for responses.